common: Fix bug in scratchpad partitioning for multiple invocations.
When we have multiple invocations of a function, we may see the base
address of a logical array cahnge. Although this information was being
updated in user_config.h, it was not getting propagated to the
LogicalArray objects, and therefore subsequent loads/stores to those
arrays could trigger assertion failures.

Change-Id: I3fde4dd15c6efdeb8fa068f54237c227a430a894