gem5: Check for overlapping array address ranges.
One common error occurs when two arrays have overlapping ranges, where
the base address + size of one array extends into the range of the
second. When this happens on a DMA or ACP access, the wrong array can be
accessed, leading to other errors or incorrect behavior. This commit
detects this condition and prints a warning message for each pair of
overlapping arrays.

Change-Id: I489039af3f7c43c33e5215c5309ba3667a040f51