unit-test: Improve manual test runner.
"make test" will now print out which test is being run and the result of
the test, while sending all of the stdout to /dev/null, so that we don't
get swamped by all the messages produced by every graph opt pass. This
makes it easier to see which tests passed and failed. Once a test fails,
the command quits immediately.

Change-Id: I06f68676beae473c141b4a9e2b25bba912503a66