common: Don't abort when an array's base address is zero.
Although this is usually not the intended behavior, there are cases
where an array that exists on an accelerator is never accessed (for
example, if an accelerator is used multiple times, there may be
certain invocations where it is unused). Thus, don't abort the
simulation, but issue a warning to the user.

Change-Id: I3434b36d2fccdbc5c54b1fc3589b1d3e2d3ac07e